+++ changelog.md
@@ -20,6 +20,8 @@
   - Option to have Favorited mods not pinned at the top of the mod list.
   - If you use Fast Rendering, its custom vmparams file (`fr.vmparams`) is now modified when you change RAM from within TriOS.
   - Option to show only disabled mods on the Dashboard mod list.
+  - New fallback method of checking if game is running that uses `wmic` if the main method (which requires a system JDK (not JRE) fails.
+  - If "Check If game is running" has errors, they appear on Settings page.
 - Fixed
   - Ship Viewer filters not working.
   - Custom launcher files are no longer hardcoded to use the game folder as their working directory.